ai-service/requirements.txt
fastapi
uvicorn
spacy
scikit-learn
pymupdf
python-docx
sentence-transformers
python-multipart
